refactor(portfolio): 🔨 Address review refinements
